+++ cpp/MTB/test/test_creator.cpp
@@ -9,10 +9,10 @@ typedef Tensor<float> TensorF;
 void test_zeros() {
   START_TEST();
   TensorF t1 = mtb::zeros<float>({2, 3});
-  if (!compare_vectors(t1.shape(), std::vector<int>{2, 3})) {
+  if (!compare_vectors(t1.shape(), std::vector<size_t>{2, 3})) {
     throw std::runtime_error("Error: t1 shape does not match expected shape after zeros!");
   }
-  for (int i = 0; i < t1.size(); ++i) {
+  for (size_t i = 0; i < t1.size(); ++i) {
     if (t1.data().get()[i] != 0.0f) {
       throw std::runtime_error("Error: t1 data is not all zeros!");
     }
@@ -24,10 +24,10 @@ void test_zeros() {
 void test_ones() {
   START_TEST();
   TensorF t1 = mtb::ones<float>({2, 3});
-  if (!compare_vectors(t1.shape(), std::vector<int>{2, 3})) {
+  if (!compare_vectors(t1.shape(), std::vector<size_t>{2, 3})) {
     throw std::runtime_error("Error: t1 shape does not match expected shape after ones!");
   }
-  for (int i = 0; i < t1.size(); ++i) {
+  for (size_t i = 0; i < t1.size(); ++i) {
     if (t1.data().get()[i] != 1.0f) {
       throw std::runtime_error("Error: t1 data is not all ones!");
     }
@@ -39,10 +39,10 @@ void test_ones() {
 void test_random() {
   START_TEST();
   TensorF t1 = mtb::random<float>({2, 3});
-  if (!compare_vectors(t1.shape(), std::vector<int>{2, 3})) {
+  if (!compare_vectors(t1.shape(), std::vector<size_t>{2, 3})) {
     throw std::runtime_error("Error: t1 shape does not match expected shape after random!");
   }
-  for (int i = 0; i < t1.size(); ++i) {
+  for (size_t i = 0; i < t1.size(); ++i) {
     if (t1.data().get()[i] < 0.0f || t1.data().get()[i] > 1.0f) {
       throw std::runtime_error("Error: t1 data is not in the range [0, 1]!");
     }
@@ -54,8 +54,8 @@ void test_random() {
 void test_triu() {
   START_TEST();
   TensorF t1({4, 4});
-  for (int i = 0; i < t1.shape()[0]; ++i) {
-    for (int j = 0; j < t1.shape()[1]; ++j) {
+  for (size_t i = 0; i < t1.shape()[0]; ++i) {
+    for (size_t j = 0; j < t1.shape()[1]; ++j) {
       t1(i, j) = static_cast<float>(i * t1.shape()[1] + j);
     }
   }
@@ -63,8 +63,8 @@ void test_triu() {
   if (!compare_vectors(t2.shape(), t1.shape())) {
     throw std::runtime_error("Error: t2 shape does not match t1 shape after triu!");
   }
-  for (int i = 0; i < t2.shape()[0]; ++i) {
-    for (int j = 0; j < t2.shape()[1]; ++j) {
+  for (size_t i = 0; i < t2.shape()[0]; ++i) {
+    for (size_t j = 0; j < t2.shape()[1]; ++j) {
       if (i > j && t2(i, j) != 0.0f) {
         throw std::runtime_error("Error: t2(" + 
           std::to_string(i) +
